requirements.txt
ultralytics==8.2.79
opencv-python>=4.9.0.80
numpy>=1.26.4
scipy>=1.11.4
matplotlib>=3.8.0
plotly>=5.18.0
pandas>=2.1.2
Flask>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

